requirements.txt
web3==6.0.0
python-dotenv==1.0.0
eth-account==0.8.0
eth-utils==2.1.0
hexbytes==0.3.0